refactor(BA-5689): move max_concurrent_logins_limit fixture into test class and improve news fragment
