+++ changes/10997.test.md
@@ -1,1 +1,1 @@
-Add unit tests for auth error HTTP status codes
+Add component test verifying that exceeding `max_concurrent_logins` returns HTTP 409 Conflict
